Merge in MSCVM/vod-cvm-voucher-partner from bugfix/CVML-6674-incorrect-attribute-in-the-incoming-response-batch-list-api to release/CVML-6490-realease-v2-voucher-partner-12-api-s-but-with-only-logic-for-my-voucher-validation

* commit 'ade8a5f718759ee841d55daf727acedf1d685abb':
  Version to 1.3.3-SNAPSHOT
  fixed null time error, renaming response object variable type
  Version to 1.3.2-SNAPSHOT
  Version to 1.3.1-SNAPSHOT
  Updated response Object renaming
  Version to 1.3.74-SNAPSHOT

+++ src/main/java/za/co/vodacom/cvm/service/impl/VPBatchServiceImpl.java
@@ -94,7 +94,7 @@ public class VPBatchServiceImpl implements VPBatchService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<BatchDetailsDTO> getVoucherQuantity(Long id, ZonedDateTime sysdate) {
+    public List<BatchDetailsDTO> getVoucherQuantity(Integer id, ZonedDateTime sysdate) {
         log.info("Request to get Voucher Quantity");
         log.debug("Request get Voucher Quantity: {}", id);
         return vPBatchRepository.getVoucherQuantity(id, sysdate);
